make type conversion for strict comparison

+++ src/Engine.php
@@ -16,7 +16,7 @@ function runGame(string $description, array $gameData): void
 
     foreach ($gameData as [$question, $correctAnswer]) {
         line("Question: {$question}");
-        $userAnswer = prompt("Your answer");
+        $userAnswer = (int) prompt("Your answer");
 
         if ($userAnswer !== $correctAnswer) {
             // phpcs:ignore
